requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.22
openai>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
markitdown[pdf,docx]>=0.1.5
slowapi>=0.1.9
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
